Move tests from init to test_*.py
